add class Cart
