+++ cart.html
@@ -92,6 +92,6 @@
       </section>
     </main>
   </body>
-  <script src="./dist/cart.js" type="module"></script>
+  <script src="./dist/cart/cart.js" type="module"></script>
 </html>
 `
