+++ index.html
@@ -459,6 +459,6 @@
         </div>
       </section>
     </footer>
-    <script src="./dist/home.js" type="module"></script>
+    <script src="./dist/product/product.js" type="module"></script>
   </body>
 </html>
