+++ typescript/constant/index.ts
@@ -1,0 +1,12 @@
+export const enum API_URL {
+  PRODUCT = './data.json',
+}
+
+export const enum LOCAL_STORAGE_KEY {
+  CART = 'cart',
+}
+
+export const enum PRODUCT_STATUS {
+  AVAILABLE = 'available',
+  OUT_OF_STOCK = 'out_of_stock',
+}
