+++ typescript/product/product.entity.ts
@@ -1,11 +1,5 @@
-export interface ProductProps {
-  id: number;
-  name: string;
-  image: string;
-  price: number;
-  discount?: number;
-  status: 'out_of_stock' | 'available';
-}
+import { PRODUCT_STATUS } from '../constant/index.js';
+import { ProductProps } from './product.interface.js';
 
 export class Product implements ProductProps {
   id: number;
@@ -13,7 +7,9 @@ export class Product implements ProductProps {
   image: string;
   price: number;
   discount: number;
-  status: 'out_of_stock' | 'available';
+  status: PRODUCT_STATUS.AVAILABLE | PRODUCT_STATUS.OUT_OF_STOCK;
+
+  discountedPrice: number;
 
   constructor({ id, name, image, price, discount, status }: ProductProps) {
     this.id = id;
@@ -22,5 +18,7 @@ export class Product implements ProductProps {
     this.price = price;
     this.discount = discount;
     this.status = status;
+
+    this.discountedPrice = (this.price * (100 - this.discount)) / 100;
   }
 }
